requirements.txt
# Django
Django==4.2.19

# Django extensions
dj-database-url==2.3.0
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-filter==24.3
django-ipware==7.0.1
django-mail-templated==2.6.5
django-modelsdoc==0.1.11
django-split-settings==1.3.2
django-unused-media==0.2.2

# Django REST Framework
djangorestframework==3.15.2
djangorestframework-filters==1.0.0.dev2
djangorestframework-simplejwt==5.4.0

# Django REST framework extensions
django-extra-fields==3.0.2
djoser==2.3.1
drf-extra-fields==3.7.0
drf-rw-serializers==1.4.0
drf-writable-nested==0.7.1
drf-yasg==1.21.8

# Python libraries
beautifulsoup4==4.13.3
# celery==5.4.0
celery==4.4.7
coverage==7.6.12
datedelta==1.4
dict2xml==1.7.6
freezegun==1.5.1
langdetect==1.0.9
lxml==5.3.1
meilisearch==0.33.1
mypy==1.15.0
psycopg2==2.9.10
pypdf==5.3.0
pysolr==3.10.0
python-dateutil==2.9.0
requests==2.32.3
redis==5.2.1
whitenoise==6.9.0
